Dump code that ultimately could not be used
This code was thought to fix the problem of making
the current digas_id an option when fetching Digas shows
failed. However, the widget approach failed because
Django validates whether it is an approved option, which
occurs outside of the widget's control. Another approach
involving the ModelForm had to be implemented instead.

This might be useful for something sooner or later, so
I'm saving this here.